Removed RGWAdmin.create_bucket
Closes #33 and #27

The `create_bucket` function called an endpoint that wasn't part of the
Admin Ops API, and whose output was incompatible with the rest of this
library as a result.  This caused any error coming back from
`create_bucket` to raise a `ServerDown`.  After some discussion in #33,
it was agreed upon that removing the feature was the best approach.

Removing this wasn't as clean as it could've been, as tests relied on it
in a few places.  I replicated the function in the test module for use
during unit tests, so as to not disrupt what the tests were already
doing.  I'm not sure this was the right approach - feedback is welcome.

**This is a breaking change**.

I did not run the test suite as I don't have access to a gateway to run
it against at present.