Use README.md as the long_description for setuptools.
This requires creating a MANIFEST.in to include the README file in
source distributions.

Markdown support on PyPi project descriptions is fairly new, so we must
install recent versions of setuptools to set the content type correctly
and have it include in PKG-INFO.

We now are also starting to build wheel packages for py3.